pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = "Auth service for The Tree Project"
authors = ["SigningOwl <whddms1208@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.7"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.37"
authlib = "^1.4.1"
pytest = "^8.3.4"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
